Move globals to App
