+++ packages/near-fast-auth-signer/src/components/InitializeGlobals/InitializeGlobals.tsx
@@ -1,11 +1,12 @@
 import { useEffect } from 'react';
 
-import FastAuthController from '../lib/controller';
-import FirestoreController from '../lib/firestoreController';
-import { networkId } from '../utils/config';
+import FastAuthController from '../../lib/controller';
+import FirestoreController from '../../lib/firestoreController';
+import { networkId } from '../../utils/config';
 
+const TEST_ACCOUNT_ID = 'harisvalj.testnet';
 window.fastAuthController = new FastAuthController({
-  accountId: 'harisvalj.testnet',
+  accountId: TEST_ACCOUNT_ID,
   networkId
 });
 
@@ -15,10 +16,13 @@ function InitializeGlobals() {
       if (!window.firestoreController) {
         window.firestoreController = new FirestoreController();
       }
-      const accountId = await window.firestoreController.getAccountIdFromOidcToken();
-
-      console.log('accountId ', accountId);
-      if (accountId) window.fastAuthController.setAccountId(accountId);
+      try {
+        const accountId = await window.firestoreController.getAccountIdFromOidcToken();
+        console.log('AccountId set as ', accountId);
+        if (accountId) window.fastAuthController.setAccountId(accountId);
+      } catch (e) {
+        console.log('AccountId not set, error: ', e.message);
+      }
     }());
   }, []); return null;
 }
